fix(gateway): align vector_metadata PK across DBs
