+++ packages/gateway/tests/unit/session-dal.postgres-list.test.ts
@@ -1,146 +1,12 @@
 import { describe, expect, it } from "vitest";
-import { dirname, join } from "node:path";
-import { fileURLToPath } from "node:url";
-import { DataType, newDb } from "pg-mem";
-import { migratePostgres } from "../../src/migrate-postgres.js";
-import type { SqlDb } from "../../src/statestore/types.js";
 import { SessionDal } from "../../src/modules/agent/session-dal.js";
 import { ChannelThreadDal } from "../../src/modules/channels/thread-dal.js";
 import { IdentityScopeDal } from "../../src/modules/identity/scope.js";
-
-const __dirname = dirname(fileURLToPath(import.meta.url));
-const postgresMigrationsDir = join(__dirname, "../../migrations/postgres");
-
-function translatePlaceholders(sql: string): { sql: string; count: number } {
-  let count = 0;
-  let out = "";
-  let inSingle = false;
-  let inDouble = false;
-
-  for (const ch of sql) {
-    if (ch === `'` && !inDouble) {
-      inSingle = !inSingle;
-      out += ch;
-      continue;
-    }
-
-    if (ch === `"` && !inSingle) {
-      inDouble = !inDouble;
-      out += ch;
-      continue;
-    }
-
-    if (ch === "?" && !inSingle && !inDouble) {
-      count += 1;
-      out += `$${count}`;
-      continue;
-    }
-
-    out += ch;
-  }
-
-  return { sql: out, count };
-}
-
-async function openPgMemDb(): Promise<{ db: SqlDb; close: () => Promise<void> }> {
-  const mem = newDb();
-
-  mem.public.registerFunction({
-    name: "jsonb_array_length",
-    args: [DataType.jsonb],
-    returns: DataType.integer,
-    implementation: (value: unknown) => {
-      if (!Array.isArray(value)) {
-        throw new Error("cannot get array length of a scalar/object");
-      }
-      return value.length;
-    },
-  });
-
-  mem.public.registerFunction({
-    name: "jsonb_typeof",
-    args: [DataType.jsonb],
-    returns: DataType.text,
-    implementation: (value: unknown) => {
-      if (value === null) return "null";
-      if (Array.isArray(value)) return "array";
-      if (typeof value === "object") return "object";
-      if (typeof value === "string") return "string";
-      if (typeof value === "number") return "number";
-      if (typeof value === "boolean") return "boolean";
-      return "unknown";
-    },
-  });
-
-  mem.public.registerFunction({
-    name: "pg_input_is_valid",
-    args: [DataType.text, DataType.text],
-    returns: DataType.bool,
-    implementation: (value: string, targetType: string) => {
-      if (!targetType || !targetType.toLowerCase().includes("json")) return false;
-      try {
-        JSON.parse(value);
-        return true;
-      } catch {
-        return false;
-      }
-    },
-  });
-
-  const { Client } = mem.adapters.createPg();
-  const pg = new Client();
-  await pg.connect();
-  await migratePostgres(pg, postgresMigrationsDir);
-
-  const db: SqlDb = {
-    kind: "postgres",
-    get: async (sql, params = []) => {
-      const translated = translatePlaceholders(sql);
-      const res = await pg.query(translated.sql, params as unknown[]);
-      return (res.rows[0] as unknown) ?? undefined;
-    },
-    all: async (sql, params = []) => {
-      const translated = translatePlaceholders(sql);
-      const res = await pg.query(translated.sql, params as unknown[]);
-      return res.rows as unknown[];
-    },
-    run: async (sql, params = []) => {
-      const translated = translatePlaceholders(sql);
-      const res = await pg.query(translated.sql, params as unknown[]);
-      return { changes: res.rowCount ?? 0 };
-    },
-    exec: async (sql) => {
-      await pg.query(sql);
-    },
-    transaction: async (fn) => {
-      await pg.query("BEGIN");
-      try {
-        const result = await fn(db);
-        await pg.query("COMMIT");
-        return result;
-      } catch (err) {
-        try {
-          await pg.query("ROLLBACK");
-        } catch {
-          // ignore rollback errors; surface original failure
-        }
-        throw err;
-      }
-    },
-    close: async () => {},
-  };
-
-  return {
-    db,
-    close: async () => {
-      await pg.end();
-    },
-  };
-}
+import { openTestPostgresDb } from "../helpers/postgres-db.js";
 
 describe("SessionDal.list (postgres)", () => {
   it("treats malformed turns_json as empty instead of failing the whole query", async () => {
-    const { db, close } = await openPgMemDb();
+    const { db, close } = await openTestPostgresDb();
     try {
       const identityScopeDal = new IdentityScopeDal(db, { cacheTtlMs: 60_000 });
       const channelThreadDal = new ChannelThreadDal(db);
@@ -176,7 +42,7 @@ describe("SessionDal.list (postgres)", () => {
   });
 
   it("treats non-array turns_json as empty instead of failing the whole query", async () => {
-    const { db, close } = await openPgMemDb();
+    const { db, close } = await openTestPostgresDb();
     try {
       const identityScopeDal = new IdentityScopeDal(db, { cacheTtlMs: 60_000 });
       const channelThreadDal = new ChannelThreadDal(db);
